% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_data.r
\name{postprocess_data}
\alias{postprocess_data}
\title{Postprocess data}
\usage{
postprocess_data(data)
}
\arguments{
\item{data}{}
}
\value{
data frame or tibble
}
\description{
A function to perform custom post-processing on data after test and training split
}
