% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.r
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data}
\usage{
preprocess_data(data)
}
\arguments{
\item{data}{}
}
\value{
data frame or tibble
}
\description{
A function to perform custom pre-processing on raw data.
}
