% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{test}
\alias{test}
\title{Run predictions from a model object}
\usage{
test(model, newdata)
}
\arguments{
\item{model}{a trained model}

\item{newdata}{new data such as a test or validation set}
}
\value{
data frame or tibble
}
\description{
A function to return model prediction on new data
}
