% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Fit Water Temperature Model}
\usage{
train(data)
}
\arguments{
\item{data}{input data frame of water temperature readings see details.}
}
\value{
GAM model object
}
\description{
This is the model that resulted from all our important research...
}
\details{
Model dataset is described in sydney_water_temp
}
